requirements.txt
asgiref==3.2.7
astroid==2.4.1
Django==3.0.7
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==7.1.2
pkg-resources==0.0.0
pylint==2.5.2
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
